perf: Lazy loading of authenticated routes
